backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
pillow
pillow-heif
piexif
python-multipart
reverse_geocoder
